Add files via upload
A dialogue system is a computer program that supports spoken, text-based, or multimodal conversational interactions with humans. Generally, a distinction is made between task-oriented and non-task-oriented dialogue systems. In task-oriented dialogues, the human user and the system engage in an interaction to accomplish some task. The system initiates a dialogue with the user  to help plan a trip and system asks the user a series of questions to obtain the information required to complete the task